Improve body armour base type prop valuation.
This commit adds prop values for dragon scale and TLA base resists.
It also adds a normalized EV prop value for all body armour types.

The EV calc here is very quick and dirty. It's certainly wrong in-game,
but it's hopefully good enough most of the time for the immediate purpose of
improving body armour valuation.