Add branch and god autotravel goals.
This commit adds inter-level branch autotravel behavior, based upon
gamestate tracking. It currently does not handle shafts or portals.

The current goalset takes the character through a 3-rune game:
lairbranch-1, lairbranch-2, vaults, and the orbrun.

During earlygame, it takes the first valid altar found, based on
the target_gods_table.

A lot of this code is pretty rough and needs further work, but it's
good enough for the moment.