Try to reduce duplicate nonmagical armour autopickup.
Crawl's ident flags are problematic. We can't be sure that an item's reported
ego or plus values are accurate until the fully identified flag has been set,
so this leads to a lot of weird handling code for that. This commit still
doesn't handle duplicate dragon armour, and there are probably other edge
cases being missed as well."