Remove an unused inventory find function.
For a function like this to be useful in crawl lua, it needs to check the
item's class. At present, ff.rc has several specialized functions to handle
this. It might be worth revisiting/readding this later, to generalize all
such instances in the code, but for now this is unused, so I'm removing it.