Adjust AUTO_DELAY options
This commit renames the old ONLINE_PLAY options, and adjusts the
default per-turn delay value to 25ms. The old (server-friendly)
delay value of 100ms is a little slow for offline play; similarly,
disabling this delay entirely tends to drop the player into tactical
situations with no context.

The Crawl devteam and server operators have requested in the past that
large Lua scripts use a 100ms delay between turns for online play, so
do be careful to adjust this if you actually use this script online.