Fix display of Low HP kickback
At some point, the Low HP kickback started not to display in the message
log until after the player pressed a key. I haven't tested to find out
if this is a recent Crawl thing or a lingering old issue with the script,
but issuing a CMD_NO_CMD_DEFAULT for each kickback *should* fix it for now.