it.fully_identified fixups
With floor-id being a thing in Crawl, many of this script's old
item.fully_identified checks became unintended interference for
the autopickup logic.

This commit removes several it.fully_identified checks,
but not all of them: Some remaining fully_identified checks
seem to be needed still, for now, to handle things like
greedy_explore autopick travel, and antique shops.

The existing state of some of this code is probably buggy,
and it will probably need another pass.