Add amulet autoequip, adjust armour / jewellery autodrop
In this commit, I've bumped the base AC weighting of body armour up somewhat,
and added a handler for dynamic jewellery autodrop.

In the future, it probably needs a dynamic check for character resistances,
so that it could try to keep things like "at least one source of rF+",
which there are currently no hard checks forcing it to do.

This code has a tendency to pick up amulets and keep them over spare rings,
as it treats all jewellery the same. I'd like to separate out that amulet
and ring handling a little bit.

This additionally fixes a you.can_train_skill() issue with autoskilling,
that I found while testing autoequip and was too lazy to separate out
into another commit.