Add pathfinding chokepoint detection
This commit contains some WIP code from the last time I was working on
ff (3 years ago); it looks like this code was intended to change the
script's target pathfinding position.

Instead of looking for a min-los position, we now look for a position
with minimum adjacent traversable tiles. As mentioned this code looks
to have been very WIP when written: Unfortunately, it wasn't committed,
and so I've lost whatever context was involved. I'm committing it now
to get it out of the way as I refamiliarize myself with this code --
probably this will need further changes.